Break things by making iteration int a long long.

+++ dynein_struct.h
@@ -324,14 +324,14 @@ typedef union {
 
 void store_onebound_PEs_callback(void* dyn, State s, void* job_msg, data_union* job_data, int iteration);
 
-void get_onebound_PE_correlation_function(generic_data* tau_data, onebound_data* corr_data, int d_tau_iter, int iterations, int max_tau_iter);
-void get_onebound_equipartition_ratio_per_runtime(generic_data* runtime_data, onebound_data* eq_data, int d_runtime_iter, int min_runtime_iter, int max_runtime_iter);
-void get_onebound_equipartition_ratio_average_per_runtime(generic_data* runtime_data, onebound_data* eq_data, int d_runtime_iter, int min_runtime_iter, int max_runtime_iter);
+void get_onebound_PE_correlation_function(generic_data* tau_data, onebound_data* corr_data, long long d_tau_iter, long long iterations, long long max_tau_iter);
+void get_onebound_equipartition_ratio_per_runtime(generic_data* runtime_data, onebound_data* eq_data, long long d_runtime_iter, long long min_runtime_iter, long long max_runtime_iter);
+void get_onebound_equipartition_ratio_average_per_runtime(generic_data* runtime_data, onebound_data* eq_data, long long d_runtime_iter, long long min_runtime_iter, long long max_runtime_iter);
 
 void print_data_to_file(double* data1, double* data2, int iterations, const char* legend, const char* fname);
 
 void simulate(double runtime, double rand_seed, State init_state, double* init_position,
-	      void (*job)(void* dyn, State s, void* job_msg, data_union* job_data, int iteration),
-	      void* job_msg, data_union* job_data);
+      void (*job)(void* dyn, State s, void* job_msg, data_union* job_data, long long iteration),
+      void* job_msg, data_union* job_data);
 
 #endif
